Goodbye decimals, hello doubles - also, new migration

+++ TaskHouseApi/Migrations/PostgresContextModelSnapshot.cs
@@ -128,15 +128,15 @@ namespace TaskHouseApi.Migrations
 
                     b.Property<bool>("Accepted");
 
-                    b.Property<int>("Complexity");
+                    b.Property<double>("Complexity");
 
-                    b.Property<decimal>("HourlyWage");
+                    b.Property<double>("HourlyWage");
 
                     b.Property<int>("TaskId");
 
                     b.Property<int>("TotalHours");
 
-                    b.Property<decimal>("Urgency");
+                    b.Property<double>("Urgency");
 
                     b.Property<int>("WorkerId");
 
@@ -285,7 +285,7 @@ namespace TaskHouseApi.Migrations
                     b.Property<int>("Id")
                         .ValueGeneratedOnAdd();
 
-                    b.Property<decimal>("AverageEstimate");
+                    b.Property<double>("AverageEstimate");
 
                     b.Property<DateTime>("Deadline");
 
@@ -299,7 +299,7 @@ namespace TaskHouseApi.Migrations
                     b.Property<string>("Title")
                         .IsRequired();
 
-                    b.Property<decimal>("Urgency");
+                    b.Property<double>("Urgency");
 
                     b.Property<string>("UrgencyString")
                         .IsRequired();
